Pass "-c" to gcc for the Dotd_Case tests.
This fixes a problem where this test was failing on some systems
due to link errors when compiling C++ source files with "gcc foo.cpp".
To link C++ code, you now need to use "g++ foo.cpp" or "gcc foo.cpp -lstdc++",
even if the C++ code in question doesn't use any C++ library features.
Rather than changing the code to use "g++" or "-lstdc++", I though it was
better to avoid the link step entirely, since it is not needed for the things
this test is testing, and removing it will speed up the test.